+++ src/main/java/com/coupang/marketplace/product/domain/Product.java
@@ -8,10 +8,10 @@ import java.math.BigInteger;
 @Getter
 public class Product {
 
-    private long id;
+    private BigInteger id;
     private BigInteger categoryId;
     private String name;
-    private long price;
+    private BigInteger price;
     private String mainImg;
     private String detailImg;
     private BigInteger stock;
@@ -22,7 +22,7 @@ public class Product {
     private Boolean rocketGlobal;
 
     @Builder
-    public Product(long id, BigInteger categoryId, String name, long price, String mainImg, String detailImg, BigInteger stock, Float score, BigInteger deliveryFee, Boolean rocket, Boolean rocketFresh, Boolean rocketGlobal) {
+    public Product(BigInteger id, BigInteger categoryId, String name, BigInteger price, String mainImg, String detailImg, BigInteger stock, Float score, BigInteger deliveryFee, Boolean rocket, Boolean rocketFresh, Boolean rocketGlobal) {
         this.id = id;
         this.categoryId = categoryId;
         this.name = name;
@@ -36,4 +36,5 @@ public class Product {
         this.rocketFresh = rocketFresh;
         this.rocketGlobal = rocketGlobal;
     }
+
 }
